added sntp time sync for timestamp_ms

+++ main/shared_data.h
@@ -11,7 +11,7 @@ typedef struct {
     uint16_t eco2;
     uint16_t tvoc;
     bool valid;
-    uint32_t timestamp_ms;
+    uint64_t timestamp_ms;
 } sensor_reading_t;
 
 typedef struct {
@@ -27,7 +27,7 @@ typedef struct {
     bool alert_eco2_bad;
     bool alert_tvoc_warn;
     bool alert_tvoc_bad;
-    uint32_t timestamp_ms;
+    uint64_t timestamp_ms;
 } processed_reading_t;
 
 extern QueueHandle_t g_sensor_queue;     // sensor_task → processing_task
